requirements.txt
cryptography==37.0.2
Django==4.0.6
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
python-dotenv==0.20
psycopg2-binary==2.9.3
geoserver-rest==2.3.4
